Fix audio delay bug

+++ SoBadItsGood/Assets/Scripts/Audio/NotePlayer.cs
@@ -14,17 +14,18 @@ public class NotePlayer : MonoBehaviour {
     /// </summary>
     /// <param name="note">The note to play</param>
     /// <param name="octave">The octave to raise it (base is middle c)</param>
-    private static int NoteToPitch(Note note, int octave) {
+    private static float NoteToPitch(Note note, int octave) {
         var octaveSemitones = octave * 12;
 
         // Formula for raising middle c by any amount of semitones: scale^semitones where scale is the 12th root of 2
         var pitch = Mathf.Pow(1.0594630943592952645618252949463f, octaveSemitones + (int)note);
 
-        return (int)pitch;
+        return pitch;
     }
 
     public void PlaySound(Note note, int octave) {
-        int pitch = NoteToPitch(note, octave);
+        float pitch = NoteToPitch(note, octave);
+        
         _src.pitch = pitch;
         _src.Play();
     }
